requirements.txt
requests
python-dotenv
jsonlines